+++ lib/clixon/clixon_xml.h
@@ -235,6 +235,7 @@ int       xml_creator_add(cxobj *xn, char *name);
 int       xml_creator_rm(cxobj *xn, char *name);
 int       xml_creator_find(cxobj *xn, char *name);
 size_t    xml_creator_len(cxobj *xn);
+cvec     *xml_creator_get(cxobj *xn);
 int       xml_creator_copy_one(cxobj *x0, cxobj *x1);
 int       xml_creator_copy_all(cxobj *x0, cxobj *x1);
 int       xml_creator_print(FILE *f, cxobj *xn);
@@ -304,7 +305,7 @@ cxobj    *xml_root(cxobj *xn);
 int       xml_operation(char *opstr, enum operation_type *op);
 char     *xml_operation2str(enum operation_type op);
 int       xml_attr_insert2val(char *instr, enum insert_type *ins);
-int       xml_add_attr(cxobj *xn, char *name, char *value, char *prefix, char *ns);
+cxobj    *xml_add_attr(cxobj *xn, char *name, char *value, char *prefix, char *ns);
 int       clicon_log_xml(int level, cxobj *x, const char *format, ...)  __attribute__ ((format (printf, 3, 4)));
 int       clixon_debug_xml(int dbglevel, cxobj *x, const char *format, ...)  __attribute__ ((format (printf, 3, 4)));